Refactor Block constructors to use time() directly

+++ src/block.cpp
@@ -2,8 +2,6 @@
 
 #include <sstream>
 #include <ctime>
-#include <string>
-#include <cstdint>
 
 // Default constructor
 Block::Block()
@@ -11,7 +9,7 @@ Block::Block()
       data(""),
       difficulty(0),
       minerAddress(""),
-      timestamp(static_cast<uint64_t>(time(nullptr))),
+      timestamp(time(nullptr)),
       nonce(0),
       reward(0),
       hash("") {}
@@ -19,13 +17,13 @@ Block::Block()
 // Parameterized constructor
 Block::Block(const std::string& prevHash,
              const std::string& blockData,
-             uint32_t diff,
+             unsigned int diff,
              const std::string& minerAddr)
     : previousHash(prevHash),
       data(blockData),
       difficulty(diff),
       minerAddress(minerAddr),
-      timestamp(static_cast<uint64_t>(time(nullptr))),
+      timestamp(time(nullptr)),
       nonce(0),
       reward(0),
       hash("") {}
